requirements.txt
certifi==2019.9.11
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
six==1.13.0
